+++ common/src/main/java/com/fej1fun/potentials/fluid/ItemFluidStorage.java
@@ -1,15 +1,12 @@
 package com.fej1fun.potentials.fluid;
 
-import com.fej1fun.potentials.capabilities.Capabilities;
 import com.fej1fun.potentials.components.FluidAmountMapDataComponent;
 import dev.architectury.fluid.FluidStack;
 import net.minecraft.core.component.DataComponentType;
-import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.material.Fluid;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;