+++ changelog.md
@@ -2,3 +2,6 @@
 
 ## Change
 - revert adding data component registry (you now need to provide your owns, like in 1.21-1.21.4)
+
+## Fix
+- fix data components are not synced to client